requirements.txt
autopep8==1.5.7
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycodestyle==2.7.0
python-dotenv==0.17.1
toml==0.10.2
Werkzeug==1.0.1
